requirements.txt
asgiref==3.8.1
Django==4.2.16
sqlparse==0.5.1
typing_extensions==4.12.2
django-cors-headers
djangorestframework
oauthlib==3.2.2
requests==2.31.0
requests-oauthlib==1.4.0
social-auth-app-django==5.4.0
social-auth-core==4.5.3
django-environ==0.10.0
djangorestframework-simplejwt==5.3.1
psycopg2-binary==2.9.9
django-money==3.5.3
djoser==2.2.3
gunicorn
drf-yasg==1.21.8